coursework_graduation (master)/lssmapper/NN_GMM_photoz/requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: linux-64
_libgcc_mutex=0.1=main
_openmp_mutex=5.1=1_gnu
abseil-cpp=20211102.0=h27087fc_1
arrow-cpp=8.0.0=py310h3098874_0
astropy=5.1=py310ha9d4c09_0
asttokens=2.0.8=pypi_0
attrs=21.4.0=pyhd3eb1b0_0
aws-c-common=0.4.57=he6710b0_1
aws-c-event-stream=0.1.6=h2531618_5
aws-checksums=0.1.9=he6710b0_0
aws-sdk-cpp=1.8.185=hce553d0_0
backcall=0.2.0=pyhd3eb1b0_0
blas=1.0=mkl
boost-cpp=1.70.0=ha2d47e9_1
bottleneck=1.3.5=py310ha9d4c09_0
brotli=1.0.9=h5eee18b_7
brotli-bin=1.0.9=h5eee18b_7
brotlipy=0.7.0=py310h7f8727e_1002
bzip2=1.0.8=h7b6447c_0
c-ares=1.18.1=h7f98852_0
ca-certificates=2022.12.7=ha878542_0
certifi=2022.12.7=pyhd8ed1ab_0
cffi=1.15.1=py310h74dc2b5_0
charset-normalizer=2.0.4=pyhd3eb1b0_0
colorama=0.4.5=py310h06a4308_0
cryptography=38.0.1=py310h9ce1e76_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=11.8.86=0
cuda-documentation=11.8.86=0
cuda-driver-dev=11.7.99=0
cuda-gdb=11.8.86=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight=11.8.86=0
cuda-nsight-compute=11.8.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=11.8.86=0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=11.8.87=0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=11.8.87=0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=11.8.86=0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pyhd3eb1b0_0
dbus=1.13.18=hb2f20db_0
debugpy=1.6.3=pypi_0
decorator=5.1.1=pyhd3eb1b0_0
einops=0.6.0=pyhd8ed1ab_0
entrypoints=0.4=py310h06a4308_0
executing=0.10.0=pypi_0
expat=2.4.9=h6a678d5_0
fast-histogram=0.10=py310h96516ba_1
fastparquet=0.5.0=py310ha9d4c09_2
ffmpeg=4.3=hf484d3e_0
fftw=3.3.9=h27cfd23_1
fontconfig=2.13.1=h6c09931_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=h4a9f257_0
gds-tools=1.4.0.31=0
gflags=2.2.2=he1b5a44_1004
giflib=5.2.1=h7b6447c_0
glib=2.69.1=h4ff587b_1
glog=0.6.0=h6f12383_0
gmp=6.2.1=h295c915_3
gnutls=3.6.15=he1e5248_0
grpc-cpp=1.46.1=h33aed49_1
gst-plugins-base=1.14.0=h8213a91_2
gstreamer=1.14.0=h28cd5cc_2
icu=58.2=he6710b0_3
idna=3.4=py310h06a4308_0
intel-openmp=2021.4.0=h06a4308_3561
ipykernel=6.15.1=pypi_0
ipython=8.4.0=py310h06a4308_0
ipywidgets=8.0.2=pyhd8ed1ab_1
jedi=0.18.1=py310h06a4308_1
joblib=1.1.1=py310h06a4308_0
jpeg=9e=h7f8727e_0
jsonschema=4.16.0=py310h06a4308_0
jupyter-client=7.3.4=pypi_0
jupyter_client=7.3.5=py310h06a4308_0
jupyter_core=4.11.1=py310h06a4308_0
jupyterlab_widgets=3.0.3=pyhd8ed1ab_0
kiwisolver=1.4.2=py310h295c915_0
krb5=1.19.2=hac12032_0
lame=3.100=h7b6447c_0
lcms2=2.12=h3be6417_0
ld_impl_linux-64=2.38=h1181459_1
lerc=3.0=h295c915_0
libblas=3.9.0=12_linux64_mkl
libbrotlicommon=1.0.9=h5eee18b_7
libbrotlidec=1.0.9=h5eee18b_7
libbrotlienc=1.0.9=h5eee18b_7
libcblas=3.9.0=12_linux64_mkl
libclang=10.0.1=default_hb85057a_2
libcublas=11.11.3.6=0
libcublas-dev=11.11.3.6=0
libcufft=10.9.0.58=0
libcufft-dev=10.9.0.58=0
libcufile=1.4.0.31=0
libcufile-dev=1.4.0.31=0
libcurand=10.3.0.86=0
libcurand-dev=10.3.0.86=0
libcurl=7.87.0=h91b91d3_0
libcusolver=11.4.1.48=0
libcusolver-dev=11.4.1.48=0
libcusparse=11.7.5.86=0
libcusparse-dev=11.7.5.86=0
libdeflate=1.8=h7f8727e_5
libedit=3.1.20210910=h7f8727e_0
libev=4.33=h516909a_1
libevent=2.1.12=h8f2d780_0
libffi=3.3=he6710b0_2
libgcc-ng=11.2.0=h1234567_1
libgfortran-ng=11.2.0=h00389a5_1
libgfortran5=11.2.0=h1234567_1
libgomp=11.2.0=h1234567_1
libiconv=1.16=h7f8727e_2
libidn2=2.3.2=h7f8727e_0
liblapack=3.9.0=12_linux64_mkl
libllvm10=10.0.1=hbcb73fb_5
libllvm11=11.1.0=hf817b99_2
libnghttp2=1.46.0=hce63b2e_0
libnpp=11.8.0.86=0
libnpp-dev=11.8.0.86=0
libnvjpeg=11.9.0.86=0
libnvjpeg-dev=11.9.0.86=0
libpng=1.6.37=hbc83047_0
libpq=12.9=h16c4e8d_3
libprotobuf=3.20.3=he621ea3_0
libsodium=1.0.18=h7b6447c_0
libssh2=1.10.0=ha56f1ee_2
libstdcxx-ng=11.2.0=h1234567_1
libtasn1=4.16.0=h27cfd23_0
libthrift=0.15.0=hcc01f38_0
libtiff=4.4.0=hecacb30_1
libunistring=0.9.10=h27cfd23_0
libuuid=1.0.3=h7f8727e_2
libwebp=1.2.4=h11a3e52_0
libwebp-base=1.2.4=h5eee18b_0
libxcb=1.15=h7f8727e_0
libxkbcommon=1.0.1=hfa300c1_0
libxml2=2.9.14=h74e7548_0
libxslt=1.1.35=h4e12654_0
lightgbm=3.3.2=py310h122e73d_0
llvmlite=0.39.1=py310he621ea3_0
lz4-c=1.9.3=h295c915_1
matplotlib=3.5.1=py310h06a4308_1
matplotlib-base=3.5.1=py310ha18d171_1
matplotlib-inline=0.1.6=py310h06a4308_0
mkl=2021.4.0=h06a4308_640
mkl-service=2.4.0=py310h7f8727e_0
mkl_fft=1.3.1=py310hd6ae3a3_0
mkl_random=1.2.2=py310h00e6091_0
mpl-scatter-density=0.7=py_0
munkres=1.1.4=py_0
nbformat=5.5.0=py310h06a4308_0
ncurses=6.3=h5eee18b_3
nest-asyncio=1.5.5=py310h06a4308_0
nettle=3.7.3=hbbd107a_1
nsight-compute=2022.3.0.22=0
nspr=4.33=h295c915_0
nss=3.74=h0370c37_0
numba=0.56.4=py310h1128e8f_0
numexpr=2.8.3=py310hcea2de6_0
numpy=1.21.6=py310h45f3432_0
openh264=2.1.1=h4ff587b_0
openssl=1.1.1t=h7f8727e_0
orc=1.7.4=h07ed6aa_0
packaging=21.3=pyhd3eb1b0_0
pandas=1.4.3=py310h6a678d5_0
parso=0.8.3=pyhd3eb1b0_0
pcre=8.45=h295c915_0
pexpect=4.8.0=pyhd3eb1b0_3
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.2.0=py310hace64e9_1
pip=22.2.2=py310h06a4308_0
ply=3.11=py310h06a4308_0
prompt-toolkit=3.0.30=pypi_0
psutil=5.9.1=pypi_0
ptyprocess=0.7.0=pyhd3eb1b0_2
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=8.0.0=py310h468efa6_0
pycparser=2.21=pyhd3eb1b0_0
pyerfa=2.0.0=py310h7f8727e_0
pygments=2.11.2=pyhd3eb1b0_0
pyopenssl=22.0.0=pyhd3eb1b0_0
pyparsing=3.0.9=py310h06a4308_0
pyqt=5.15.7=py310h6a678d5_1
pyqt5-sip=12.11.0=pypi_0
pyrsistent=0.18.0=py310h7f8727e_0
pysocks=1.7.1=py310h06a4308_0
python=3.10.6=haa1d7c7_1
python-dateutil=2.8.2=pyhd3eb1b0_0
python-fastjsonschema=2.16.2=py310h06a4308_0
python_abi=3.10=2_cp310
pytorch=1.13.0=py3.10_cuda11.7_cudnn8.5.0_0
pytorch-cuda=11.7=h67b0de4_0
pytorch-mutex=1.0=cuda
pytz=2022.1=py310h06a4308_0
pyyaml=6.0=pypi_0
pyzmq=23.2.1=pypi_0
qt-main=5.15.2=h327a75a_7
qt-webengine=5.15.9=hd2b0992_4
qtwebkit=5.212=h4eab89a_4
re2=2022.04.01=h27087fc_0
readline=8.2=h5eee18b_0
requests=2.28.1=py310h06a4308_0
scikit-learn=1.1.1=py310h6a678d5_0
scipy=1.9.1=py310hd5efca6_0
setuptools=65.5.0=py310h06a4308_0
sip=6.6.2=py310h6a678d5_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=h295c915_0
sqlite=3.39.3=h5082296_0
stack-data=0.4.0=pypi_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.7.0=hdb19cb5_0
threadpoolctl=2.2.0=pyh0d69192_0
thrift=0.17.0=py310h6a678d5_0
tk=8.6.12=h1ccaba5_0
toml=0.10.2=pyhd3eb1b0_0
torchaudio=0.13.0=py310_cu117
torchvision=0.14.0=py310_cu117
tornado=6.2=py310h5eee18b_0
tqdm=4.64.0=pyhd8ed1ab_0
traitlets=5.3.0=pypi_0
typing_extensions=4.3.0=py310h06a4308_0
tzdata=2022f=h04d1e81_0
urllib3=1.26.12=py310h06a4308_0
utf8proc=2.6.1=h27cfd23_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.37.1=pyhd3eb1b0_0
widgetsnbextension=4.0.3=pyhd8ed1ab_0
xz=5.2.6=h5eee18b_0
yaml=0.2.5=h7b6447c_0
zeromq=4.3.4=h2531618_0
zlib=1.2.13=h5eee18b_0
zstd=1.5.2=ha4553b6_0
